requirements.txt
fastapi==0.115.0       
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.34
alembic==1.13.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic==2.9.2
pydantic-settings==2.4.0
email-validator==2.2.0
bcrypt==4.2.0
httpx==0.28.1
